Updating reading, testing links
Used two paragraphs from the exec summary for the README.

I *think* links might not want any prepending. (E.g. no / on local
links.) For local testing, the / is a problem, as it suggests an
absolute path rooted at the root of the domain. With no slash, it
suggests a file in the same directory/at the same path level as the
document, allowing a local renderer to correctly render the link.

Pushing to test.